[fix]: validator のパスが違っていたので、修正しました 　→ model 配下に置いていましたが、 Rails Way 的には validators 配下におくべきだと考えているので、修正しました
